src/requirements.txt
beautifulsoup4==4.14.3
blinker==1.9.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
dash==4.1.0
dnspython==2.8.0
Flask==3.1.3
idna==3.11
importlib_metadata==9.0.0
interchange==2021.0.4
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==6.1.0
MarkupSafe==3.0.3
monotonic==1.6
narwhals==2.19.0
neo4j==6.1.0
nest-asyncio==1.6.0
numpy==2.4.4
packaging==26.1
pandas==3.0.2
pansi==2024.11.0
pillow==12.2.0
plotly==6.7.0
psycopg2-binary==2.9.11
py2neo==2021.2.4
Pygments==2.20.0
pymongo==4.16.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
requests==2.33.1
retrying==1.4.2
six==1.17.0
soupsieve==2.8.3
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
Werkzeug==3.1.8
zipp==3.23.1
apache-airflow==2.5.0
apache-airflow-providers-apache-kafka==2.0.0
kafka-python==2.0.2